+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Added icons to game and map editor executables on Windows
 * Added ServerSettings.DisableWhenEmpty config for servers
 * Log time is now more precise and will always reflect elapsed time (before that change, log time have been the same during each update loop)
+* Decreased server usage (it will now sleep according to its tick rate)
 
 ### Fixes